Move BookStats model to models.
